requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
addict=2.2.0=py36_0
albumentations=0.3.0=py36_0
asn1crypto=0.24.0=py36_3
backcall=0.1.0=py36_0
blas=1.0=mkl
bleach=3.0.2=py36_0
ca-certificates=2019.8.28=0
certifi=2019.9.11=py36_0
cffi=1.11.5=py36h74b6da3_1
chardet=3.0.4=py36_3
click=7.0=py36_0
cloudpickle=0.5.5=py36_0
colorama=0.4.0=py36_0
cryptography=2.4.2=py36h7a1dbc1_0
cryptography-vectors=2.3.1=py36_0
cuda80=1.0=0
cudatoolkit=9.2=0
cycler=0.10.0=py36h009560c_0
dask-core=0.19.0=py36_0
decorator=4.3.0=py36_0
dominate=2.3.5=pypi_0
entrypoints=0.2.3=py36_2
flask=1.0.2=py36_1
freetype=2.9.1=ha9979f8_1
geos=3.7.2=habb2df7_1
gettext=0.19.8.1=hb01d8f6_1002
glib=2.58.3=hc0c2ac7_1002
hdf5=1.8.20=hac2f561_1
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
idna=2.7=py36_2
imageio=2.3.0=py36_0
imgaug=0.2.9=py_1
intel-openmp=2018.0.3=0
ipykernel=5.1.0=py36h39e3cac_0
ipython=7.1.1=py36h39e3cac_0
ipython_genutils=0.2.0=py36h3c5d0ee_0
itsdangerous=1.1.0=py36_0
jedi=0.13.1=py36_0
jinja2=2.10=py36_0
joblib=0.13.2=py36_0
jpeg=9b=hb83a4c4_2
jsonschema=2.6.0=py36h7636477_0
jupyter-contrib-core=0.3.3=pypi_0
jupyter-contrib-nbextensions=0.5.1=pypi_0
jupyter-highlight-selected-word=0.2.0=pypi_0
jupyter-latex-envs=1.4.6=pypi_0
jupyter-nbextensions-configurator=0.4.1=pypi_0
jupyter_client=5.2.3=py36_0
jupyter_core=4.4.0=py36_0
jupyterlab=0.35.3=py36_0
jupyterlab_server=0.2.0=py36_0
kiwisolver=1.0.1=py36h6538335_0
libffi=3.2.1=h6538335_1006
libiconv=1.15=hfa6e2cd_1005
libopencv=3.4.2=h20b85fd_0
libpng=1.6.35=h2a8f88b_0
libsodium=1.0.16=h9d3ae62_0
libtiff=4.0.9=h36446d0_2
libwebp=0.5.2=vc14_7
lxml=4.4.1=pypi_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=1.0=py36hfa6e2cd_1
matplotlib=3.0.2=py36hc8f65d3_0
mistune=0.8.3=py36hfa6e2cd_1
mkl=2019.4=245
mkl-service=2.0.2=py36he774522_0
mkl_fft=1.0.14=py36h14836fe_0
mkl_random=1.0.2=py36h343c172_0
msys2-conda-epoch=20160418=1
munch=2.3.2=pypi_0
nbconvert=5.3.1=py36_0
nbformat=4.4.0=py36h3a5bc1b_0
networkx=2.1=py36_0
ninja=1.9.0=py36h74a9793_0
notebook=5.6.0=py36_0
numpy=1.15.4=py36h19fb1c0_0
numpy-base=1.15.4=py36hc3f5095_0
olefile=0.45.1=py36_0
opencv=3.4.2=py36h40b0b35_0
openssl=1.1.1d=he774522_2
pandas=0.25.0=py36ha925a31_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py36_1
parso=0.3.1=py36_0
pcre=8.41=h6538335_1003
pickleshare=0.7.5=py36_0
pillow=5.3.0=py36hdc69c19_0
pip=18.1=pypi_0
pkg-config=0.29.2=h8cfccfd_1005
pretrainedmodels=0.7.4=pypi_0
prometheus_client=0.4.2=py36_0
prompt_toolkit=2.0.7=py36_0
py-opencv=3.4.2=py36hc319ecb_0
pybind11=2.4.2=py36h74a9793_0
pycparser=2.18=py36_1
pygments=2.2.0=py36hb010967_0
pyopenssl=18.0.0=py36_0
pyparsing=2.2.0=py36_1
pyqt=5.9.2=py36h6538335_2
pyqt5=5.13.0=pypi_0
pyqt5-sip=4.19.18=pypi_0
pysocks=1.6.8=py36_2
python=3.6.6=hea74fb7_0
python-dateutil=2.7.3=py36_0
pytorch=1.2.0=py3.6_cuda92_cudnn7_1
pytz=2018.5=py36_0
pywavelets=0.5.2=py36h452e1ab_2
pywinpty=0.5.4=py36_0
pyyaml=3.13=pypi_0
pyzmq=17.1.2=py36hf576995_0
qt=5.9.7=vc14h73c81de_0
qt5=5.4.0=0
requests=2.19.1=py36_1
scikit-image=0.14.0=py36h6538335_1
scikit-learn=0.21.2=py36h6288b17_0
scipy=1.3.1=py36h29ff71c_0
segmentation-models-pytorch=0.0.1=pypi_0
send2trash=1.5.0=py36_0
setuptools=40.2.0=py36_0
shapely=1.6.4=py36ha35856d_1006
sip=4.19.8=py36h6538335_0
six=1.11.0=py36_1
sqlite=3.26.0=he774522_0
terminado=0.8.1=py36_1
testpath=0.4.2=py36_0
tk=8.6.8=hfa6e2cd_0
toolz=0.9.0=py36_0
torchfile=0.1.0=py_0
torchnet=0.0.4=pypi_0
torchvision=0.2.1=pypi_0
tornado=5.1=py36hfa6e2cd_0
tqdm=4.25.0=py36h28b3542_0
traitlets=4.3.2=py36h096827d_0
urllib3=1.23=py36_1
vc=14.1=h0510ff6_4
visdom=0.1.8.5=0
vs2015_runtime=14.16.27012=hf0eaf9b_0
vs2017_runtime=15.4.27004.2010=1
wcwidth=0.1.7=py36h3d5aa90_0
webencodings=0.5.1=py36_1
websocket-client=0.52.0=py36_0
werkzeug=0.14.1=py36_0
wheel=0.31.1=py36_0
win_inet_pton=1.0.1=py36_2
wincertstore=0.2=py36h7fe50ca_0
winpty=0.4.3=4
yaml=0.1.7=vc14_0
zeromq=4.2.5=he025d50_1
zlib=1.2.11=vc14_0
